initial audit service refactoring
