+++ activiti-cloud-starter-audit-mongo/src/main/java/org/activiti/cloud/starter/audit/configuration/ActivitiAuditMongoAutoConfiguration.java
@@ -1,0 +1,10 @@
+package org.activiti.cloud.starter.audit.configuration;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("org.activiti.cloud.services.audit.mongo")
+public class ActivitiAuditMongoAutoConfiguration {
+
+}
