+++ activiti-cloud-starter-audit-mongo/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    org.activiti.cloud.starter.audit.configuration.ActivitiAuditMongoAutoConfiguration
