+++ activiti-cloud-starter-audit-mongo/src/test/resources/application.properties
@@ -15,8 +15,6 @@ keycloak.realm=springboot
 keycloak.resource=activiti
 activiti.keycloak.test-user=testuser
 activiti.keycloak.test-password=password
-spring.cloud.discovery.enabled=false
-eureka.client.enabled=false
 
 spring.data.mongodb.host=localhost
 spring.data.mongodb.port=27017
