+++ activiti-cloud-starter-audit/src/main/java/org/activiti/cloud/starter/audit/configuration/ActivitiAuditAutoConfiguration.java
@@ -1,0 +1,10 @@
+package org.activiti.cloud.starter.audit.configuration;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("org.activiti.cloud.services.audit")
+public class ActivitiAuditAutoConfiguration {
+
+}
