+++ activiti-cloud-starter-audit/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    org.activiti.cloud.starter.audit.configuration.ActivitiAuditAutoConfiguration
